Supported languages

Language	Code	Examples

Afrikaans	af	1 en 2	1, 2 en 3	1 of 2	1, 2 of 3

Amharic	am	1 እና 2	1፣ 2, እና 3	1 ወይም 2﻿	1፣ 2፣ ወይም 3

Arabic	ar	1 و2	1 و2 و3	1 أو 2	1 أو 2 أو 3

Assamese	as	1 আৰু 2	1, 2 আৰু 3	1 বা 2	1, 2 বা 3

Azerbaijani	az	1 və 2	1, 2 və 3	1 yaxud 2	1, 2, yaxud 3

Belarusian	be	1 і 2	1, 2 і 3	1 ці 2	1, 2 ці 3

Bulgarian	bg	1 и 2	1, 2 и 3	1 или 2	1, 2 или 3

Bangla	bn	1 এবং 2	1, 2 এবং 3	1 বা 2	1, 2, বা 3

Bosnian	bs	1 i 2	1, 2 i 3	1 ili 2	1, 2 ili 3

Catalan	ca	1 i 2	1, 2 i 3	1 o 2	1, 2 o 3

Czech	cs	1 a 2	1, 2 a 3	1 nebo 2	1, 2 nebo 3

Welsh	cy	1 a 2	1, 2 a 3	1 neu 2	1, 2 neu 3

Danish	da	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

German	de	1 und 2	1, 2 und 3	1 oder 2	1, 2 oder 3

Greek	el	1 και 2	1, 2 και 3	1 ή 2	1, 2 ή 3

English	en	1 and 2	1, 2 and 3	1 or 2	1, 2 or 3

English (United Arab Emirates)	en_AE	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (American Samoa)	en_AS	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Burundi)	en_BI	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Guam)	en_GU	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Marshall Islands)	en_MH	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Northern Mariana Islands)	en_MP	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Philippines)	en_PH	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Puerto Rico)	en_PR	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (U.S. Outlying Islands)	en_UM	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (United States)	en_US	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (U.S. Virgin Islands)	en_VI	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

Spanish	es	1 y 2	1, 2 y 3	1 o 2	1, 2 o 3

Estonian	et	1 ja 2	1, 2 ja 3	1 või 2	1, 2 või 3

Basque	eu	1 eta 2	1, 2 eta 3	1 edo 2	1, 2 edo 3

Persian	fa	1 و 2	1،‏ 2، و 3	1 یا 2	1،‏ 2، یا 3

Finnish	fi	1 ja 2	1, 2 ja 3	1 tai 2	1, 2 tai 3

Filipino	fil	1 at 2	1, 2, at 3	1 o 2	1, 2, o 3

French	fr	1 et 2	1, 2 et 3	1 ou 2	1, 2 ou 3

Irish	ga	1 agus 2	1, 2 agus 3	1 nó 2	1, 2, nó 3

Galician	gl	1 e 2	1, 2 e 3	1 ou 2	1, 2 ou 3

Gujarati	gu	1 અને 2	1, 2 અને 3	1 અથવા 2	1, 2, અથવા 3

Hebrew	he	1 ו2	1, 2 ו3	1 או 2	1, 2, או 3

Hindi	hi	1 और 2	1, 2, और 3	1 या 2	1, 2 या 3

Croatian	hr	1 i 2	1, 2 i 3	1 ili 2	1, 2 ili 3

Hungarian	hu	1 és 2	1, 2 és 3	1 vagy 2	1, 2 vagy 3

Armenian	hy	1 և 2	1, 2 և 3	1 կամ 2	1, 2 կամ 3

Indonesian	id	1 dan 2	1, 2, dan 3	1 atau 2	1, 2, atau 3

Icelandic	is	1 og 2	1, 2 og 3	1 eða 2	1, 2 eða 3

Italian	it	1 e 2	1, 2 e 3	1 o 2	1, 2 o 3

Japanese	ja	1、2	1、2、3	1または2	1、2、または3

Javanese	jv	1 lan 2	1, 2, lan 3	1 utowo 2	1, 2, utowo 3

Georgian	ka	1 და 2	1, 2 და 3	1 ან 2	1, 2 ან 3

Kazakh	kk	1 және 2	1, 2, 3	1 не 2	1, 2, не болмаса 3

Khmer	km	1 និង​2	1, 2 និង 3	1 ឬ 2	1, 2 ឬ 3

Kannada	kn	1 ಮತ್ತು 2	1, 2, ಮತ್ತು 3	1 ಅಥವಾ 2	1, 2, ಅಥವಾ 3

Korean	ko	1 및 2	1, 2 및 3	1 또는 2	1, 2 또는 3

Kyrgyz	ky	1 жана 2	1, 2 жана 3	1 же 2	1, 2 же 3

Lao	lo	1 ແລະ 2	1, 2, 3	1 ຫຼື 2	1, 2 ຫຼື 3

Lithuanian	lt	1 ir 2	1, 2 ir 3	1 ar 2	1, 2 ar 3

Latvian	lv	1 un 2	1, 2 un 3	1 vai 2	1, 2 vai 3

Macedonian	mk	1 и 2	1, 2 и 3	1 или 2	1, 2 или 3

Malayalam	ml	1 കൂടാതെ 2	1, 2, 3 എന്നിവ	1 അല്ലെങ്കിൽ 2	1, 2, അല്ലെങ്കിൽ 3

Mongolian	mn	1, 2	1, 2, 3	1 эсвэл 2	1, 2, 3 зэргийн аль нэг

Marathi	mr	1 आणि 2	1, 2 आणि 3	1 किंवा 2	1, 2, किंवा 3

Malay	ms	1 dan 2	1, 2 dan 3	1 atau 2	1, 2, atau 3

Burmese	my	1နှင့် 2	1 - 2နှင့် 3	1 သို့မဟုတ် 2	1 - 2 သို့မဟုတ် 3

Norwegian Bokmål	nb	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Nepali	ne	1 र 2	1,2 र 3	1 वा 2	1, 2, वा 3

Dutch	nl	1 en 2	1, 2 en 3	1 of 2	1, 2 of 3

Norwegian Nynorsk	nn	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Norwegian	no	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Odia	or	1 ଓ 2	1, 2, ଓ 3	1 କିମ୍ବା 2	1, 2 କିମ୍ବା 3

Punjabi	pa	1 ਅਤੇ 2	1, 2 ਅਤੇ 3	1 ਜਾਂ 2	1, 2 ਜਾਂ 3

Polish	pl	1 i 2	1, 2 i 3	1 lub 2	1, 2 lub 3

Pashto	ps	1 او 2	1، 2، او 3	1 or 2	1, 2, یا 3

Portuguese	pt	1 e 2	1, 2 e 3	1 ou 2	1, 2 ou 3

Romanian	ro	1 și 2	1, 2 și 3	1 sau 2	1, 2 sau 3

Russian	ru	1 и 2	1, 2 и 3	1 или 2	1, 2 или 3

Sindhi	sd	1 ۽ 2	1, 2، ۽ 3	1 يا 2	1, 2, يا 3

Sinhala	si	1 සහ 2	1, 2, සහ 3	1 හෝ 2	1, 2, හෝ 3

Slovak	sk	1 a 2	1, 2 a 3	1 alebo 2	1, 2 alebo 3

Slovenian	sl	1 in 2	1, 2 in 3	1 ali 2	1, 2 ali 3

Somali	so	1 iyo 2	1, 2 iyo 3	1 ama 2	1, 2 ama 3

Albanian	sq	1 dhe 2	1, 2 dhe 3	1 ose 2	1, 2 ose 3

Serbian	sr	1 и 2	1, 2 и 3	1 или 2	1, 2 или 3

Swedish	sv	1 och 2	1, 2 och 3	1 eller 2	1, 2 eller 3

Swahili	sw	1 na 2	1, 2 na 3	1 au 2	1, 2 au 3

Tamil	ta	1 மற்றும் 2	1, 2 மற்றும் 3	1 அல்லது 2	1, 2 அல்லது 3

Telugu	te	1 మరియు 2	1, 2 మరియు 3	1 లేదా 2	1, 2 లేదా 3

Thai	th	1และ2	1 2 และ3	1 หรือ 2	1, 2 หรือ 3

Turkmen	tk	1 we 2	1, 2 we 3	1 ýa-da 2	1, 2 ýa-da 3

Turkish	tr	1 ve 2	1, 2 ve 3	1 veya 2	1, 2 veya 3

Ukrainian	uk	1 і 2	1, 2 і 3	1 або 2	1, 2 або 3

Urdu	ur	1 اور 2	1، 2، اور 3	1 یا 2	1، 2، یا 3

Uzbek	uz	1 va 2	1, 2 va 3	1 yoki 2	1, 2 yoki 3

Vietnamese	vi	1 và 2	1, 2 và 3	1 hoặc 2	1, 2 hoặc 3

Cantonese	yue	1同2	1、2同3	1 或 2	1、2 或 3

Chinese	zh	1和2	1、2和3	1或2	1、2或3

Zulu	zu	1 ne-2	1, 2, ne-3	1 or 2	1, 2, or 3

Usage

and_languages

Format

A data frame with 3 variables:

language

The name of the language, possibly with a territory in parentheses

code

The language code

example_and_2

An example of a conjunctive list with two elements in the language

example_and_3

An example of a conjunctive list with three elements in the language

example_or_2

An example of a disjunctive list with two elements in the language

example_or_3

An example of a disjunctive list with three elements in the language

Source

Language data is derived from the Unicode Common Locale Data Repository (CLDR)