Supported languages

Language	Code	Examples

Afrikaans	af	1 en 2	1, 2 en 3	1 of 2	1, 2 of 3

Amharic	am	1 <U+12A5><U+1293> 2	1<U+1363> 2, <U+12A5><U+1293> 3	1 <U+12C8><U+12ED><U+121D> 2<U+FEFF>	1<U+1363> 2<U+1363> <U+12C8><U+12ED><U+121D> 3

Arabic	ar	1 <U+0648>2	1 <U+0648>2 <U+0648>3	1 <U+0623><U+0648> 2	1 <U+0623><U+0648> 2 <U+0623><U+0648> 3

Assamese	as	1 <U+0986><U+09F0><U+09C1> 2	1, 2 <U+0986><U+09F0><U+09C1> 3	1 <U+09AC><U+09BE> 2	1, 2 <U+09AC><U+09BE> 3

Azerbaijani	az	1 v<U+0259> 2	1, 2 v<U+0259> 3	1 yaxud 2	1, 2, yaxud 3

Belarusian	be	1 <U+0456> 2	1, 2 <U+0456> 3	1 <U+0446><U+0456> 2	1, 2 <U+0446><U+0456> 3

Bulgarian	bg	1 <U+0438> 2	1, 2 <U+0438> 3	1 <U+0438><U+043B><U+0438> 2	1, 2 <U+0438><U+043B><U+0438> 3

Bangla	bn	1 <U+098F><U+09AC><U+0982> 2	1, 2 <U+098F><U+09AC><U+0982> 3	1 <U+09AC><U+09BE> 2	1, 2, <U+09AC><U+09BE> 3

Bosnian	bs	1 i 2	1, 2 i 3	1 ili 2	1, 2 ili 3

Catalan	ca	1 i 2	1, 2 i 3	1 o 2	1, 2 o 3

Czech	cs	1 a 2	1, 2 a 3	1 nebo 2	1, 2 nebo 3

Welsh	cy	1 a 2	1, 2 a 3	1 neu 2	1, 2 neu 3

Danish	da	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

German	de	1 und 2	1, 2 und 3	1 oder 2	1, 2 oder 3

Greek	el	1 <U+03BA>a<U+03B9> 2	1, 2 <U+03BA>a<U+03B9> 3	1 <U+03AE> 2	1, 2 <U+03AE> 3

English	en	1 and 2	1, 2 and 3	1 or 2	1, 2 or 3

English (United Arab Emirates)	en_AE	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (American Samoa)	en_AS	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Burundi)	en_BI	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Guam)	en_GU	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Marshall Islands)	en_MH	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Northern Mariana Islands)	en_MP	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Philippines)	en_PH	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (Puerto Rico)	en_PR	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (U.S. Outlying Islands)	en_UM	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (United States)	en_US	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

English (U.S. Virgin Islands)	en_VI	1 and 2	1, 2, and 3	1 or 2	1, 2, or 3

Spanish	es	1 y 2	1, 2 y 3	1 o 2	1, 2 o 3

Estonian	et	1 ja 2	1, 2 ja 3	1 või 2	1, 2 või 3

Basque	eu	1 eta 2	1, 2 eta 3	1 edo 2	1, 2 edo 3

Persian	fa	1 <U+0648> 2	1<U+060C><U+200F> 2<U+060C> <U+0648> 3	1 <U+06CC><U+0627> 2	1<U+060C><U+200F> 2<U+060C> <U+06CC><U+0627> 3

Finnish	fi	1 ja 2	1, 2 ja 3	1 tai 2	1, 2 tai 3

Filipino	fil	1 at 2	1, 2, at 3	1 o 2	1, 2, o 3

French	fr	1 et 2	1, 2 et 3	1 ou 2	1, 2 ou 3

Irish	ga	1 agus 2	1, 2 agus 3	1 nó 2	1, 2, nó 3

Galician	gl	1 e 2	1, 2 e 3	1 ou 2	1, 2 ou 3

Gujarati	gu	1 <U+0A85><U+0AA8><U+0AC7> 2	1, 2 <U+0A85><U+0AA8><U+0AC7> 3	1 <U+0A85><U+0AA5><U+0AB5><U+0ABE> 2	1, 2, <U+0A85><U+0AA5><U+0AB5><U+0ABE> 3

Hebrew	he	1 <U+05D5>2	1, 2 <U+05D5>3	1 <U+05D0><U+05D5> 2	1, 2, <U+05D0><U+05D5> 3

Hindi	hi	1 <U+0914><U+0930> 2	1, 2, <U+0914><U+0930> 3	1 <U+092F><U+093E> 2	1, 2 <U+092F><U+093E> 3

Croatian	hr	1 i 2	1, 2 i 3	1 ili 2	1, 2 ili 3

Hungarian	hu	1 és 2	1, 2 és 3	1 vagy 2	1, 2 vagy 3

Armenian	hy	1 <U+0587> 2	1, 2 <U+0587> 3	1 <U+056F><U+0561><U+0574> 2	1, 2 <U+056F><U+0561><U+0574> 3

Indonesian	id	1 dan 2	1, 2, dan 3	1 atau 2	1, 2, atau 3

Icelandic	is	1 og 2	1, 2 og 3	1 eða 2	1, 2 eða 3

Italian	it	1 e 2	1, 2 e 3	1 o 2	1, 2 o 3

Japanese	ja	1<U+3001>2	1<U+3001>2<U+3001>3	1<U+307E><U+305F><U+306F>2	1<U+3001>2<U+3001><U+307E><U+305F><U+306F>3

Javanese	jv	1 lan 2	1, 2, lan 3	1 utowo 2	1, 2, utowo 3

Georgian	ka	1 <U+10D3><U+10D0> 2	1, 2 <U+10D3><U+10D0> 3	1 <U+10D0><U+10DC> 2	1, 2 <U+10D0><U+10DC> 3

Kazakh	kk	1 <U+0436><U+04D9><U+043D><U+0435> 2	1, 2, 3	1 <U+043D><U+0435> 2	1, 2, <U+043D><U+0435> <U+0431><U+043E><U+043B><U+043C><U+0430><U+0441><U+0430> 3

Khmer	km	1 <U+1793><U+17B7><U+1784><U+200B>2	1, 2 <U+1793><U+17B7><U+1784> 3	1 <U+17AC> 2	1, 2 <U+17AC> 3

Kannada	kn	1 <U+0CAE><U+0CA4><U+0CCD><U+0CA4><U+0CC1> 2	1, 2, <U+0CAE><U+0CA4><U+0CCD><U+0CA4><U+0CC1> 3	1 <U+0C85><U+0CA5><U+0CB5><U+0CBE> 2	1, 2, <U+0C85><U+0CA5><U+0CB5><U+0CBE> 3

Korean	ko	1 <U+BC0F> 2	1, 2 <U+BC0F> 3	1 <U+B610><U+B294> 2	1, 2 <U+B610><U+B294> 3

Kyrgyz	ky	1 <U+0436><U+0430><U+043D><U+0430> 2	1, 2 <U+0436><U+0430><U+043D><U+0430> 3	1 <U+0436><U+0435> 2	1, 2 <U+0436><U+0435> 3

Lao	lo	1 <U+0EC1><U+0EA5><U+0EB0> 2	1, 2, 3	1 <U+0EAB><U+0EBC><U+0EB7> 2	1, 2 <U+0EAB><U+0EBC><U+0EB7> 3

Lithuanian	lt	1 ir 2	1, 2 ir 3	1 ar 2	1, 2 ar 3

Latvian	lv	1 un 2	1, 2 un 3	1 vai 2	1, 2 vai 3

Macedonian	mk	1 <U+0438> 2	1, 2 <U+0438> 3	1 <U+0438><U+043B><U+0438> 2	1, 2 <U+0438><U+043B><U+0438> 3

Malayalam	ml	1 <U+0D15><U+0D42><U+0D1F><U+0D3E><U+0D24><U+0D46> 2	1, 2, 3 <U+0D0E><U+0D28><U+0D4D><U+0D28><U+0D3F><U+0D35>	1 <U+0D05><U+0D32><U+0D4D><U+0D32><U+0D46><U+0D19><U+0D4D><U+0D15><U+0D3F><U+0D7D> 2	1, 2, <U+0D05><U+0D32><U+0D4D><U+0D32><U+0D46><U+0D19><U+0D4D><U+0D15><U+0D3F><U+0D7D> 3

Mongolian	mn	1, 2	1, 2, 3	1 <U+044D><U+0441><U+0432><U+044D><U+043B> 2	1, 2, 3 <U+0437><U+044D><U+0440><U+0433><U+0438><U+0439><U+043D> <U+0430><U+043B><U+044C> <U+043D><U+044D><U+0433>

Marathi	mr	1 <U+0906><U+0923><U+093F> 2	1, 2 <U+0906><U+0923><U+093F> 3	1 <U+0915><U+093F><U+0902><U+0935><U+093E> 2	1, 2, <U+0915><U+093F><U+0902><U+0935><U+093E> 3

Malay	ms	1 dan 2	1, 2 dan 3	1 atau 2	1, 2, atau 3

Burmese	my	1<U+1014><U+103E><U+1004><U+1037><U+103A> 2	1 - 2<U+1014><U+103E><U+1004><U+1037><U+103A> 3	1 <U+101E><U+102D><U+102F><U+1037><U+1019><U+101F><U+102F><U+1010><U+103A> 2	1 - 2 <U+101E><U+102D><U+102F><U+1037><U+1019><U+101F><U+102F><U+1010><U+103A> 3

Norwegian Bokmål	nb	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Nepali	ne	1 <U+0930> 2	1,2 <U+0930> 3	1 <U+0935><U+093E> 2	1, 2, <U+0935><U+093E> 3

Dutch	nl	1 en 2	1, 2 en 3	1 of 2	1, 2 of 3

Norwegian Nynorsk	nn	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Norwegian	no	1 og 2	1, 2 og 3	1 eller 2	1, 2 eller 3

Odia	or	1 <U+0B13> 2	1, 2, <U+0B13> 3	1 <U+0B15><U+0B3F><U+0B2E><U+0B4D><U+0B2C><U+0B3E> 2	1, 2 <U+0B15><U+0B3F><U+0B2E><U+0B4D><U+0B2C><U+0B3E> 3

Punjabi	pa	1 <U+0A05><U+0A24><U+0A47> 2	1, 2 <U+0A05><U+0A24><U+0A47> 3	1 <U+0A1C><U+0A3E><U+0A02> 2	1, 2 <U+0A1C><U+0A3E><U+0A02> 3

Polish	pl	1 i 2	1, 2 i 3	1 lub 2	1, 2 lub 3

Pashto	ps	1 <U+0627><U+0648> 2	1<U+060C> 2<U+060C> <U+0627><U+0648> 3	1 or 2	1, 2, <U+06CC><U+0627> 3

Portuguese	pt	1 e 2	1, 2 e 3	1 ou 2	1, 2 ou 3

Romanian	ro	1 <U+0219>i 2	1, 2 <U+0219>i 3	1 sau 2	1, 2 sau 3

Russian	ru	1 <U+0438> 2	1, 2 <U+0438> 3	1 <U+0438><U+043B><U+0438> 2	1, 2 <U+0438><U+043B><U+0438> 3

Sindhi	sd	1 <U+06FD> 2	1, 2<U+060C> <U+06FD> 3	1 <U+064A><U+0627> 2	1, 2, <U+064A><U+0627> 3

Sinhala	si	1 <U+0DC3><U+0DC4> 2	1, 2, <U+0DC3><U+0DC4> 3	1 <U+0DC4><U+0DDD> 2	1, 2, <U+0DC4><U+0DDD> 3

Slovak	sk	1 a 2	1, 2 a 3	1 alebo 2	1, 2 alebo 3

Slovenian	sl	1 in 2	1, 2 in 3	1 ali 2	1, 2 ali 3

Somali	so	1 iyo 2	1, 2 iyo 3	1 ama 2	1, 2 ama 3

Albanian	sq	1 dhe 2	1, 2 dhe 3	1 ose 2	1, 2 ose 3

Serbian	sr	1 <U+0438> 2	1, 2 <U+0438> 3	1 <U+0438><U+043B><U+0438> 2	1, 2 <U+0438><U+043B><U+0438> 3

Swedish	sv	1 och 2	1, 2 och 3	1 eller 2	1, 2 eller 3

Swahili	sw	1 na 2	1, 2 na 3	1 au 2	1, 2 au 3

Tamil	ta	1 <U+0BAE><U+0BB1><U+0BCD><U+0BB1><U+0BC1><U+0BAE><U+0BCD> 2	1, 2 <U+0BAE><U+0BB1><U+0BCD><U+0BB1><U+0BC1><U+0BAE><U+0BCD> 3	1 <U+0B85><U+0BB2><U+0BCD><U+0BB2><U+0BA4><U+0BC1> 2	1, 2 <U+0B85><U+0BB2><U+0BCD><U+0BB2><U+0BA4><U+0BC1> 3

Telugu	te	1 <U+0C2E><U+0C30><U+0C3F><U+0C2F><U+0C41> 2	1, 2 <U+0C2E><U+0C30><U+0C3F><U+0C2F><U+0C41> 3	1 <U+0C32><U+0C47><U+0C26><U+0C3E> 2	1, 2 <U+0C32><U+0C47><U+0C26><U+0C3E> 3

Thai	th	1<U+0E41><U+0E25><U+0E30>2	1 2 <U+0E41><U+0E25><U+0E30>3	1 <U+0E2B><U+0E23><U+0E37><U+0E2D> 2	1, 2 <U+0E2B><U+0E23><U+0E37><U+0E2D> 3

Turkmen	tk	1 we 2	1, 2 we 3	1 ýa-da 2	1, 2 ýa-da 3

Turkish	tr	1 ve 2	1, 2 ve 3	1 veya 2	1, 2 veya 3

Ukrainian	uk	1 <U+0456> 2	1, 2 <U+0456> 3	1 <U+0430><U+0431><U+043E> 2	1, 2 <U+0430><U+0431><U+043E> 3

Urdu	ur	1 <U+0627><U+0648><U+0631> 2	1<U+060C> 2<U+060C> <U+0627><U+0648><U+0631> 3	1 <U+06CC><U+0627> 2	1<U+060C> 2<U+060C> <U+06CC><U+0627> 3

Uzbek	uz	1 va 2	1, 2 va 3	1 yoki 2	1, 2 yoki 3

Vietnamese	vi	1 và 2	1, 2 và 3	1 ho<U+1EB7>c 2	1, 2 ho<U+1EB7>c 3

Cantonese	yue	1<U+540C>2	1<U+3001>2<U+540C>3	1 <U+6216> 2	1<U+3001>2 <U+6216> 3

Chinese	zh	1<U+548C>2	1<U+3001>2<U+548C>3	1<U+6216>2	1<U+3001>2<U+6216>3

Zulu	zu	1 ne-2	1, 2, ne-3	1 or 2	1, 2, or 3

Usage

and_languages

Format

A data frame with 3 variables:

language

The name of the language, possibly with a territory in parentheses

code

The language code

example_and_2

An example of a conjunctive list with two elements in the language

example_and_3

An example of a conjunctive list with three elements in the language

example_or_2

An example of a disjunctive list with two elements in the language

example_or_3

An example of a disjunctive list with three elements in the language

Source

Language data is derived from the Unicode Common Locale Data Repository (CLDR)