Supported languages

A list of supported languages and examples of their usage.

Usage

and_languages

Format

A data frame with 6 variables:

language

The name of the language, possibly with a territory in parentheses

code

The language code

example_and_2

An example of a conjunctive list with two elements in the language

example_and_4

An example of a conjunctive list with four elements in the language

example_or_2

An example of a disjunctive list with two elements in the language

example_or_4

An example of a disjunctive list with four elements in the language

support

Either "full" or "partial". Partially supported languages generally localize and() but not or().

Source

Language data is derived from the Unicode Common Locale Data Repository (CLDR)

Examples

and_languages
#> # A tibble: 227 × 7
#>    language  code  example_and_2 example_and_4 example_or_2 example_or_4 support
#>    <chr>     <chr> <chr>         <chr>         <chr>        <chr>        <chr>  
#>  1 Afrikaans af    1 en 2        1, 2, 3 en 4  1 of 2       1, 2, 3 of 4 full   
#>  2 Albanian  sq    1 dhe 2       1, 2, 3 dhe 4 1 ose 2      1, 2, 3 ose… full   
#>  3 Amharic   am    1 እና 2        1፣ 2፣ 3, እና 4 1 ወይም 2﻿      1፣ 2፣ 3፣ ወይ… full   
#>  4 Arabic    ar    ‏1 و2‎          ‏1 و2 و3 و4‎    ‏1 أو 2‎       1 أو 2 أو 3… full   
#>  5 Armenian  hy    1 և 2         1, 2, 3 և 4   1 կամ 2      1, 2, 3 կամ… full   
#>  6 Assamese  as    1 আৰু 2        1, 2, 3 আৰু 4  1 বা 2       1, 2, 3 বা 4 full   
#>  7 Asturian  ast   1 y 2         1, 2, 3 y 4   1 o 2        1, 2, 3 o 4  full   
#>  8 Azerbaij… az    1 və 2        1, 2, 3 və 4  1 yaxud 2    1, 2, 3, ya… full   
#>  9 Bangla    bn    1 এবং 2       1, 2, 3 এবং 4 1 বা 2       1, 2, 3, বা… full   
#> 10 Basque    eu    1 eta 2       1, 2, 3 eta 4 1 edo 2      1, 2, 3 edo… full   
#> # ℹ 217 more rows